More file suffix changes, plus some file renaming. Also fixed some broken tests (compilation, broken macros, etc.)
Made all unit test targets 'test-*' for easy lookup and consistency.
